requirements.txt
mysql-connector
numpy
imghdr
cv2